backend/requirements.txt
# AI APIs
openai>=1.0.0
requests>=2.28.0
google-generativeai>=0.8.0

# Config
pydantic>=2.0.0
pydantic-settings>=2.2.0

# Database (for full pipeline)
supabase>=2.0.0

# Entity matching
rapidfuzz>=3.6.0
